+++ pkg/webui/components/wizard/wizard.js
@@ -20,9 +20,11 @@ import renderCallback from '../../lib/render-callback'
 
 import { WizardContext } from './context'
 
+const FIRST_STEP = 1
+
+// Action types.
 const INIT = 'INIT'
 const GO_TO_STEP = 'GO_TO_STEP'
-const SET_SNAPSHOT = 'SET_SNAPSHOT'
 
 const reducer = (state, action) => {
   switch (action.type) {
@@ -32,14 +34,20 @@ const reducer = (state, action) => {
         steps: action.steps,
       }
     case GO_TO_STEP:
+      const { snapshots: oldSnapshots, currentStep: oldStep } = state
+      const { values, step: currentStep } = action
+
+      // Replace current step values when navigating between wizard steps.
+      const snapshots = [
+        ...oldSnapshots.slice(0, oldStep - 1),
+        values,
+        ...oldSnapshots.slice(oldStep),
+      ]
+
       return {
         ...state,
-        currentStep: action.step,
-      }
-    case SET_SNAPSHOT:
-      return {
-        ...state,
-        snapshot: merge({}, state.snapshot, action.snapshot),
+        currentStep,
+        snapshots,
       }
     default:
       return state
@@ -50,12 +58,16 @@ const Wizard = props => {
   const { initialStep, onComplete, initialValues, completeMessage } = props
 
   const [state, dispatch] = React.useReducer(reducer, {
+    // Active step in the wizard.
     currentStep: initialStep,
+    // A list of all steps in the wizard.
     steps: [],
-    snapshot: initialValues,
+    // A list of form values for each step in the wizard.
+    // For example, `snapshots[0]` - has form values after submitting the first step.
+    snapshots: [],
   })
 
-  const { currentStep, steps, snapshot } = state
+  const { currentStep, steps, snapshots } = state
 
   const stepsCount = steps.length
 
@@ -63,19 +75,18 @@ const Wizard = props => {
     dispatch({ type: INIT, steps })
   }, [])
   const prevStep = React.useCallback(
-    values => {
-      dispatch({ type: SET_SNAPSHOT, snapshot: values })
-      dispatch({ type: GO_TO_STEP, step: Math.min(currentStep - 1, stepsCount) })
-    },
-    [currentStep, stepsCount],
+    values => dispatch({ type: GO_TO_STEP, step: Math.max(currentStep - 1, FIRST_STEP), values }),
+    [currentStep],
   )
   const nextStep = React.useCallback(
-    values => {
-      dispatch({ type: SET_SNAPSHOT, snapshot: values })
-      dispatch({ type: GO_TO_STEP, step: Math.min(currentStep + 1, stepsCount) })
-    },
+    values => dispatch({ type: GO_TO_STEP, step: Math.min(currentStep + 1, stepsCount), values }),
     [currentStep, stepsCount],
   )
+
+  const snapshot = React.useMemo(() => merge({}, initialValues, ...snapshots), [
+    initialValues,
+    snapshots,
+  ])
 
   const context = {
     completeMessage,